api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.23
alembic==1.13.1
influxdb-client==1.38.0
websockets==12.0
asyncio-mqtt==0.16.1
httpx==0.25.2
psutil==5.9.6
pandas==2.1.4
openpyxl==3.1.2
python-dotenv==1.0.0
redis==5.0.1
celery==5.3.4
aiofiles==23.2.1
Pillow==10.1.0
pytz==2023.3